Fix dataSource, get more data about single paper
SemanticScholar API doesn't provide as much data in the profile
papers endpoint as much as in the single paper endpoint itself.
e.g., when using the profile papers endpoint the following data
is mostly missing:
hIndex for co-authors: unavailable
references/citations mostly non-existent.

for that we delegate the hundreds of get papers related get
requests to a server that acts like a proxy and delegates
the get requests to the API as is.